% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{find_variable_events}
\alias{find_variable_events}
\title{Calculate the Sum Deviance for Inclusion and Exclusion Matrices}
\usage{
find_variable_events(
  m1_matrix,
  m2_matrix,
  min_row_sum = 50,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{m1_matrix}{A matrix representing the inclusion matrix. Rows are events, columns are barcodes.}

\item{m2_matrix}{A matrix representing the exclusion matrix. Rows are events, columns are barcodes.}

\item{min_row_sum}{A numeric value specifying the minimum row sum threshold for filtering events. Defaults to 50.}

\item{verbose}{Logical. If \code{TRUE} (default), prints progress and informational messages.}

\item{...}{Additional arguments to be passed.}
}
\value{
A \code{data.table} containing the events and their corresponding sum deviance values.
}
\description{
Calculate the Sum Deviance for Inclusion and Exclusion Matrices
}
