% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{get_silhouette_mean}
\alias{get_silhouette_mean}
\title{Compute Average Silhouette Width with Logging}
\usage{
get_silhouette_mean(X, cluster_assignments, n_threads = 1, verbose = FALSE)
}
\arguments{
\item{X}{A numeric matrix where rows are observations and columns are features.}

\item{cluster_assignments}{An integer vector of cluster assignments, which must be the same length as the number of rows in \code{X}.}

\item{n_threads}{Number of threads to use for parallel processing.}

\item{verbose}{Logical. If \code{TRUE}, prints progress and informational messages. Default is \code{FALSE}.}
}
\value{
A single numeric value: the average silhouette score.
}
\description{
Computes the average silhouette width for a clustering solution using Euclidean distance.
}
\note{
This process can be very slow for large matrices if single-threaded. Use multiple threads to take advantage of parallel computation for significantly faster results.
}
\examples{
# Preparing the inputs
set.seed(42)
pc_matrix <- matrix(data = rnorm(n = 10000 * 15, sd = 2), nrow = 10000, ncol = 15)
cluster_numbers <- runif(n = 10000, min = 1, max = 10) |> as.integer()

# Getting the mean silhouette score
n_threads <- parallel::detectCores()
score <- get_silhouette_mean(pc_matrix, cluster_numbers, n_threads)
print(score)

}
