% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_solo_processing.R
\name{load_toy_M1_M2_object}
\alias{load_toy_M1_M2_object}
\title{Load the toy M1/M2 object}
\usage{
load_toy_M1_M2_object()
}
\value{
An R object from the toy_m1_m2_obj.rds file.
}
\description{
Loads a toy object of M1 and M2 used for examples or testing.
}
