% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_solo_processing.R
\name{load_toy_SJ_object}
\alias{load_toy_SJ_object}
\title{Load the toy SJ object}
\usage{
load_toy_SJ_object()
}
\value{
An R object from the toy_SJ_object.RDS file.
}
\description{
Loads a toy splice junction object used for examples or testing.
}
