% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_solo_processing.R
\name{make_m2}
\alias{make_m2}
\title{make_m2}
\usage{
make_m2(m1_inclusion_matrix, eventdata)
}
\arguments{
\item{m1_inclusion_matrix}{A sparse matrix to be modified and used for creating the M2 matrix.}

\item{eventdata}{A data.table containing event information with at least \code{group_id} and an index column.}
}
\value{
A sparse matrix M2 with the dummy row removed and proper adjustments made.
}
\description{
Creates the M2 matrix from a given m1_inclusion_matrix and eventdata, ensuring the proper processing of group indices and matrix operations.
}
\examples{
junction_abundance_object <- load_toy_SJ_object()
m1_obj <- make_m1(junction_ab_object = junction_abundance_object)

# obtaining the m1 and eventdata
m1 <- m1_obj$m1_inclusion_matrix
eventdata <- m1_obj$event_data
m2 <- make_m2(m1_inclusion_matrix = m1, eventdata = eventdata)

}
