% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_m2_cpp}
\alias{make_m2_cpp}
\title{Legacy function - redirects to memory-efficient version}
\usage{
make_m2_cpp(M1, group_ids, n_threads = 1L)
}
\arguments{
\item{M1}{Sparse matrix (dgCMatrix) of inclusion counts (events x cells)}

\item{group_ids}{Integer vector of group IDs for each event}

\item{n_threads}{Number of threads for OpenMP (default 1)}
}
\value{
Sparse matrix M2 with same dimensions as M1
}
\description{
Legacy function - redirects to memory-efficient version
}
\keyword{internal}
