% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splikit-package.R
\docType{package}
\name{splikit-package}
\alias{splikit-package}
\title{splikit: Analysing RNA Splicing in scRNA-Seq Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides analysis of high-dimensional single-cell splicing data. Offers a framework to extract and work with ratio-based data structures derived from single-cell RNA sequencing experiments. Provides both a modern R6 object-oriented interface and direct matrix manipulation functions. Core functionalities are implemented in C++ via 'Rcpp' to ensure high performance and scalability on large datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://csglab.github.io/splikit/}
  \item \url{https://github.com/csglab/splikit}
  \item Report bugs at \url{https://github.com/csglab/splikit/issues}
}

}
\author{
\strong{Maintainer}: Arsham Mikaeili Namini \email{arsham.mikaeilinamini@mail.mcgill.ca} (\href{https://orcid.org/0000-0002-9453-6951}{ORCID})

}
\keyword{internal}
