% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplikitObject.R
\name{splikit}
\alias{splikit}
\title{Create a SplikitObject}
\usage{
splikit(...)
}
\arguments{
\item{...}{Arguments passed to \code{SplikitObject$new()}.}
}
\value{
A new SplikitObject instance.
}
\description{
Convenience function to create a SplikitObject.
}
\examples{
\dontrun{
# From junction abundance
junction_ab <- load_toy_SJ_object()
obj <- splikit(junction_ab = junction_ab)

# From existing matrices
obj <- splikit(m1 = my_m1, m2 = my_m2, eventData = my_eventdata)
}

}
