% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{find_variable_events}
\alias{find_variable_events}
\title{Calculate the Sum Deviance for Inclusion and Exclusion Matrices}
\usage{
find_variable_events(
  m1_matrix,
  m2_matrix,
  min_row_sum = 50,
  verbose = TRUE,
  n_threads = 1,
  ...
)
}
\arguments{
\item{m1_matrix}{A matrix representing the inclusion matrix. Rows are events, columns are barcodes.}

\item{m2_matrix}{A matrix representing the exclusion matrix. Rows are events, columns are barcodes.}

\item{min_row_sum}{A numeric value specifying the minimum row sum threshold for filtering events. Defaults to 50.}

\item{verbose}{Logical. If \code{TRUE} (default), prints progress and informational messages.}

\item{n_threads}{If the module OpenPM is available for your device, the function suggests using multi-thread processing for even faster computation.}

\item{...}{Additional arguments to be passed.}
}
\value{
A \code{data.table} containing the events and their corresponding sum deviance values.
}
\description{
Calculate the Sum Deviance for Inclusion and Exclusion Matrices
}
\examples{
# loading the toy dataset
toy_obj <- load_toy_M1_M2_object()

# getting HVE (high variable events)
 HVE <- find_variable_events(toy_obj$m1, toy_obj$m2)

 # printing the results
 print(HVE[order(-sum_deviance)])
}
