% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.R
\name{get_pseudo_correlation}
\alias{get_pseudo_correlation}
\title{Compute Pseudo-Correlation Using Beta-Binomial Model}
\usage{
get_pseudo_correlation(
  ZDB_matrix,
  m1_inclusion,
  m2_exclusion,
  suppress_warnings = TRUE
)
}
\arguments{
\item{ZDB_matrix}{A numeric dense matrix of shape (events x samples). Should have rownames representing events.}

\item{m1_inclusion}{A numeric dense matrix of the same number of rows as \code{ZDB_matrix}, representing inclusion features.}

\item{m2_exclusion}{A numeric dense matrix of the same number of rows as \code{ZDB_matrix}, representing exclusion features.}

\item{suppress_warnings}{Logical. If \code{TRUE} (default), suppresses warnings during any warnings triggered during
computation (e.g., due to ill-conditioned inputs)}
}
\value{
A \code{data.table} with the following columns:
\describe{
\item{event}{The event names from \code{ZDB_matrix} rownames.}
\item{pseudo_correlation}{The computed pseudo R² correlation values.}
\item{null_distribution}{Null correlation values from a permuted version of \code{ZDB_matrix}.}
}
}
\description{
This function calculates a pseudo R²-like correlation metric using a beta-binomial model
implemented in C++. It takes in a data matrix \code{ZDB_matrix} and two model matrices
for inclusion and exclusion, respectively. Optionally, it can suppress warnings
generated during computation.
}
\examples{
set.seed(42)
# get the m1 object
junction_abundance_object <- load_toy_SJ_object()
m1_obj <- make_m1(junction_ab_object = junction_abundance_object)

# obtaining the m1 and eventdata
m1_inclusion <- m1_obj$m1_inclusion_matrix
eventdata <- m1_obj$event_data
m2_exclusion <- make_m2(m1_inclusion, eventdata)

# creating a dummy ZDB
ZDB_matrix <- matrix(rnorm(n = (nrow(m1_inclusion) * ncol(m1_inclusion)), sd = 7),
nrow = nrow(m1_inclusion),
ncol = ncol(m1_inclusion))
rownames(ZDB_matrix) <- rownames(m1_inclusion)

# both m1 and m2 need to be a dense matrix
m1_inclusion <- as.matrix(m1_inclusion)
m2_exclusion <- as.matrix(m2_exclusion)

# getting the pseudo-correlation
pseudo_r_squre <- get_pseudo_correlation(ZDB_matrix, m1_inclusion, m2_exclusion)
print(pseudo_r_squre)

}
