% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplikitObject.R
\name{splikit}
\alias{splikit}
\title{Create a SplikitObject}
\usage{
splikit(...)
}
\arguments{
\item{...}{Arguments passed to \code{SplikitObject$new()}.}
}
\value{
A new SplikitObject instance.
}
\description{
Convenience function to create a SplikitObject.
}
\examples{
\donttest{
# From existing matrices using the toy dataset
toy <- load_toy_M1_M2_object()
obj <- splikit(m1 = toy$m1, m2 = toy$m2, eventData = toy$eventdata)
}

}
